Move all source files into `/files/src`
