+++ files/src/modules/background.js
