+++ files/src/modules/running_as_electron.js
