+++ files/src/nibbler.html
@@ -49,32 +49,32 @@
 	module form. Still, for the time being we have everything in one namespace
 	and so load them in a particular order... -->
 
-	<script src="js_renderer/10_globals.js"></script>
-	<script src="js_renderer/20_utils.js"></script>
-	<script src="js_renderer/30_point.js"></script>
-	<script src="js_renderer/31_sliders.js"></script>
-	<script src="js_renderer/40_position.js"></script>
-	<script src="js_renderer/41_fen.js"></script>
-	<script src="js_renderer/42_perft.js"></script>
-	<script src="js_renderer/43_chess960.js"></script>
-	<script src="js_renderer/50_table.js"></script>
-	<script src="js_renderer/51_node.js"></script>
-	<script src="js_renderer/52_sorted_moves.js"></script>
-	<script src="js_renderer/55_winrate_graph.js"></script>
-	<script src="js_renderer/60_pgn_utils.js"></script>
-	<script src="js_renderer/61_pgn_parse.js"></script>
-	<script src="js_renderer/63_polyglot.js"></script>
-	<script src="js_renderer/65_loaders.js"></script>
-	<script src="js_renderer/71_tree_handler.js"></script>
-	<script src="js_renderer/72_tree_draw.js"></script>
-	<script src="js_renderer/75_looker.js"></script>
-	<script src="js_renderer/80_info.js"></script>
-	<script src="js_renderer/81_arrows.js"></script>
-	<script src="js_renderer/82_infobox.js"></script>
-	<script src="js_renderer/83_statusbox.js"></script>
-	<script src="js_renderer/90_engine.js"></script>
-	<script src="js_renderer/95_hub.js"></script>
-	<script src="js_renderer/99_start.js"></script>
+	<script src="renderer/10_globals.js"></script>
+	<script src="renderer/20_utils.js"></script>
+	<script src="renderer/30_point.js"></script>
+	<script src="renderer/31_sliders.js"></script>
+	<script src="renderer/40_position.js"></script>
+	<script src="renderer/41_fen.js"></script>
+	<script src="renderer/42_perft.js"></script>
+	<script src="renderer/43_chess960.js"></script>
+	<script src="renderer/50_table.js"></script>
+	<script src="renderer/51_node.js"></script>
+	<script src="renderer/52_sorted_moves.js"></script>
+	<script src="renderer/55_winrate_graph.js"></script>
+	<script src="renderer/60_pgn_utils.js"></script>
+	<script src="renderer/61_pgn_parse.js"></script>
+	<script src="renderer/63_polyglot.js"></script>
+	<script src="renderer/65_loaders.js"></script>
+	<script src="renderer/71_tree_handler.js"></script>
+	<script src="renderer/72_tree_draw.js"></script>
+	<script src="renderer/75_looker.js"></script>
+	<script src="renderer/80_info.js"></script>
+	<script src="renderer/81_arrows.js"></script>
+	<script src="renderer/82_infobox.js"></script>
+	<script src="renderer/83_statusbox.js"></script>
+	<script src="renderer/90_engine.js"></script>
+	<script src="renderer/95_hub.js"></script>
+	<script src="renderer/99_start.js"></script>
 
 </body>
 </html>
