+++ files/src/pieces/_B.png
