+++ files/src/renderer/31_sliders.js
