+++ files/src/renderer/80_info.js
